Rename dequant_i32 operator to dequant_i32_bf16 for naming consistency
Align operator directory name with the kernel file convention
(dequant_i32_bf16.cc). The existing dequant (int4→bf16) and quantize
operators use incomplete type suffixes; this rename establishes the
full input_output naming pattern.
